#1823 Rearrange kernel module-inline tests
