+++ src/psyclone/tests/psyir/transformations/kernel_transformation_test.py
@@ -48,7 +48,6 @@ from psyclone.domain.lfric.lfric_builtins import LFRicBuiltIn
 from psyclone.generator import GenerationError
 from psyclone.psyGen import Kern
 from psyclone.psyir.nodes import Routine, FileContainer
-from psyclone.psyir.symbols import SymbolError
 from psyclone.psyir.transformations import TransformationError
 from psyclone.transformations import ACCRoutineTrans, \
     Dynamo0p3KernelConstTrans
@@ -56,9 +55,6 @@ from psyclone.transformations import ACCRoutineTrans, \
 from psyclone.tests.gocean_build import GOceanBuild
 from psyclone.tests.lfric_build import LFRicBuild
 from psyclone.tests.utilities import get_invoke
-
-# TODO: move this file to the proper location?
-from psyclone.domain.common.transformations import KernelModuleInlineTrans
 
 
 def setup_module():
@@ -336,45 +332,3 @@ def test_builtin_no_trans():
         rtrans.apply(kernels[0])
     assert ("ACCRoutineTrans to a built-in kernel is not yet supported and "
             "kernel 'x_plus_y' is of type " in str(err.value))
-
-
-def test_no_inline_global_var():
-    ''' Check that we refuse to in-line a kernel that accesses a global
-    variable. '''
-    inline_trans = KernelModuleInlineTrans()
-    _, invoke = get_invoke("single_invoke_kern_with_global.f90",
-                           api="gocean1.0", idx=0)
-    sched = invoke.schedule
-    kernels = sched.walk(Kern)
-    with pytest.raises(TransformationError) as err:
-        inline_trans.apply(kernels[0])
-    assert ("'kernel_with_global_code' contains accesses to data (variable "
-            "'alpha') that are not present in the Symbol Table(s) "
-            "within KernelSchedule scope." in str(err.value))
-
-
-# Class KernelTrans
-
-# Method validate
-
-def test_kernel_trans_validate(monkeypatch):
-    '''Check that the validate method in the class KernelTrans raises an
-    exception if the kernel code can not be retrieved.
-
-    '''
-    kernel_trans = KernelModuleInlineTrans()
-    _, invoke = get_invoke("single_invoke_kern_with_global.f90",
-                           api="gocean1.0", idx=0)
-    sched = invoke.schedule
-    kernels = sched.walk(Kern)
-    kernel = kernels[0]
-
-    def raise_symbol_error():
-        '''Simple function that raises SymbolError.'''
-        raise SymbolError("error")
-    monkeypatch.setattr(kernel, "get_kernel_schedule", raise_symbol_error)
-    with pytest.raises(TransformationError) as err:
-        kernel_trans.apply(kernel)
-    assert ("KernelModuleInline failed to retrieve PSyIR for kernel "
-            "'kernel_with_global_code' using the 'get_kernel_schedule' "
-            "method." in str(err.value))
